PPM calibration Phase 2+3: auto carrier-offset measurement + sub-ppm resolution
Phase 3 (sub-ppm): ppm_correction_value is now a double (fractional ppm), so a
fraction of a ppm resolves to a Hz-accurate shift — needed on the high bands.
radio_ppm_correction() does full floating math (llround(f_rf*ppm/1e6)); both ppm
spinners gain 0.01 steps / 2 digits; persisted as %f (old int props load via
atof). Protocol 1 rx/tx now route through the shared helper too.

Phase 2 (auto measurement): ppm_cal gains an FFT-based carrier-offset measurement
that discovers the ppm error from a reference station and applies it closed-loop.
It works on the raw off-air I/Q (demod-independent), tapped in receiver.c via
ppm_cal_iq_feed() with an atomic fast-path no-op when idle. The active RX is
retuned so the carrier lands at a fixed audio offset (clear of the DC spike);
32768-pt Hann-windowed frames go through a self-contained radix-2 FFT (no
FT8/kiss/fftw dependency), the carrier bin is parabolic-interpolated to sub-bin,
and several frames are averaged against Doppler with an SNR gate. Correction:
ppm_new = ppm_old + (residual/(carrier-lo))*1e6.

UI: a Calibrate button in Configure -> Misc runs it on a background poll and
auto-applies the result; poll + widget refs are torn down on page destroy.

Verified: builds clean; boots under --faker with the live tap; a standalone FFT
test confirms the sign convention (+/- tone -> +/- offset) and sub-Hz accuracy
(~0.17 Hz worst case at 384 kHz). On-air verification against a real carrier
still pending.

+++ radio.h
@@ -117,7 +117,7 @@ typedef struct _radio {
   gdouble panadapter_calibration;
   gdouble swr_alarm_value;
   gint temperature_alarm_value;
-  int ppm_correction_value;
+  double ppm_correction_value;  // oscillator error, parts-per-million (fractional)
   int ppm_ref_station;     // index into the ppm_cal.c reference-station table
 
   gint cw_keyer_sidetone_frequency;
